+++ PacManGame/src/main/java/de/amr/games/pacman/controller/event/PacManGainsPowerEvent.java
@@ -2,9 +2,9 @@ package de.amr.games.pacman.controller.event;
 
 public class PacManGainsPowerEvent implements PacManGameEvent {
 
-	public final int duration;
+	public final long duration;
 
-	public PacManGainsPowerEvent(int ticks) {
+	public PacManGainsPowerEvent(long ticks) {
 		this.duration = ticks;
 	}
 }
